Create new component for menu
