+++ src/components/Header.jsx
@@ -1,29 +1,18 @@
-import { Image, StyleSheet, Text, TouchableOpacity, View } from "react-native";
+import {
+  Image,
+  StyleSheet,
+  Text,
+  TouchableOpacity,
+  View,
+  Modal,
+} from "react-native";
 import React, { useState } from "react";
+import Menu from "./Menu";
 import assets from "../constants/assets";
 import colors from "../constants/colors";
 
 const Header = () => {
   const [isModalVisible, setIsModalVisible] = useState(false);
-
-  const menuItems = [
-    {
-      text: "Share",
-      icon: assets.shareIconDarkTheme,
-    },
-    {
-      text: "Like",
-      icon: assets.likeIconDarkTheme,
-    },
-    {
-      text: "Light Mode",
-      icon: assets.lightIconDarkTheme,
-    },
-    {
-      text: "More Info",
-      icon: assets.infoIconDarkTheme,
-    },
-  ];
 
   // Function to show or hide the menu
   //-----------------------------------
@@ -51,30 +40,7 @@ const Header = () => {
       </TouchableOpacity>
 
       {/* Menu */}
-      {isModalVisible && (
-        <TouchableOpacity
-          activeOpacity={1} // To prevent extra click events
-          onPress={toggleModal}
-          style={styles.modalContainer}
-        >
-          {menuItems.map((item, index) => (
-            <View key={index}>
-              <TouchableOpacity
-                style={styles.menuItem}
-                onPress={() => console.log("menu item")}
-              >
-                <Image source={item.icon} style={styles.menuItemIcon} />
-                <Text style={styles.menuText}>{item.text}</Text>
-              </TouchableOpacity>
-
-              {/* Separator for menu items */}
-              {index < menuItems.length - 1 && (
-                <View style={styles.separator}></View>
-              )}
-            </View>
-          ))}
-        </TouchableOpacity>
-      )}
+      {isModalVisible && <Menu toggleModal={toggleModal} />}
     </View>
   );
 };
@@ -112,35 +78,5 @@ const styles = StyleSheet.create({
   menuIcon: {
     width: 40,
     height: 40,
-  },
-  modalContainer: {
-    backgroundColor: colors.menuBgColorDarkTheme,
-    position: "absolute",
-    top: "130%",
-    left: 200,
-    right: 14,
-    zIndex: 200,
-  },
-  menuItem: {
-    flexDirection: "row",
-    gap: 12,
-    paddingVertical: 15,
-    paddingHorizontal: 10,
-  },
-  menuItemIcon: {
-    width: 20,
-    height: 20,
-  },
-  menuSection: {
-    position: "relative",
-  },
-  menuText: {
-    color: colors.textColorDarkTheme,
-    fontSize: 16,
-  },
-  separator: {
-    height: 1,
-    width: "100%",
-    backgroundColor: "#707070",
   },
 });
